==> notebooks/step_1_scrape.ipynb <==
# Imports

import requests
from bs4 import BeautifulSoup
import re
from playwright.async_api import async_playwright
import pandas as pd
import time
pd.options.display.max_rows = None
pd.options.display.max_colwidth = None

#Starting up Playwright

url_hear = 'https://www.judiciary.senate.gov/hearings'

playwright = await async_playwright().start()
browser = await playwright.chromium.launch(headless=True)
page = await browser.new_page()

await page.goto(url_hear)

# Going through each nomination meeting (on the first page) and grabbing the district judges and orgs who subitted briefs in support of them.
# Strangely, when I have Playwright .click() on each link, it's as if I right clicked. Never encountered that before.

meetings = await page.query_selector_all('a[href*="meetings"]')
links = [await meeting.get_attribute('href') for meeting in meetings]
data = []

for i in range(len(links)):
    link = links[i]
    if link.find('nominations') != -1:
        full_link = 'https://www.judiciary.senate.gov' + link
        await page.goto(full_link)
        time.sleep(1)
    
        names = await page.query_selector_all('li[class=" col-xs-12 "] span.fn')
        positions = await page.query_selector_all('li[class=" col-xs-12 "] div.title')
        docs = await page.query_selector_all('li[class="acrobat"] a')

        for j in range(len(positions)):
            position = str(await positions[j].inner_text()).strip()
            name = str(await names[j].inner_text()).strip()
            last_name = name.split(' ')[-1]
            if position.find('District Judge') != -1:
                temp_dict = {
                    'name': name,
                    'position': position,
                    'docs': []
                }
                for k in range(len(docs)):
                    doc = str(await docs[k].inner_text()).strip()
                    if doc.find(last_name) != -1:
                        temp_dict['docs'].append(doc)

            data.append(temp_dict)

        await page.go_back()
        

# Going through all the pages! Well, all the pages that have suporting documents posted, which I manually checked.



all_data = []

for i in range(1, 9):
    await page.goto(url_hear + '?PageNum_rs=' + str(i))
    time.sleep(1)
    meetings = await page.query_selector_all('a[href*="meetings"]')
    links = [await meeting.get_attribute('href') for meeting in meetings]
    data = []

    for j in range(len(links)):
        link = links[j]
        if link.find('nominations') != -1:
            full_link = 'https://www.judiciary.senate.gov' + link
            await page.goto(full_link)
            time.sleep(1)

            names = await page.query_selector_all('li[class=" col-xs-12 "] span.fn')
            positions = await page.query_selector_all('li[class=" col-xs-12 "] div.title')
            docs = await page.query_selector_all('li[class="acrobat"] a')
            dates = await page.query_selector_all('tr[class="date "] td')

            try:
                date = await dates[1].inner_text()
            except:
                dates = await page.query_selector_all('tr[class="date"] td')
                date = await dates[1].inner_text()

            beginning = 'https://www.judiciary.senate.gov/'
            end = '&download=1'

            for k in range(len(positions)):
                position = str(await positions[k].inner_text()).strip()
                temp_dict={}
                
                if position.find('District Judge') != -1:
                    name = str(await names[k].inner_text()).strip()
                    last_name = name.split(' ')[-1]
                    temp_dict = {
                        'name': name,
                        'position': position.lower().replace('to be ',''),
                        'docs': [],
                        'doc_links': [],
                        'date': date
                    }
                    
                    # This is because the statements/letters/briefs in support of someone are in the format "group support for [last name]"
                    for m in range(len(docs)):
                        doc = str(await docs[m].inner_text()).strip()
                        if (doc.find(last_name) != -1) & (doc.find('For') != -1):
                            temp_dict['docs'].append(doc)
                            link_doc = doc.lower().replace(' ','-')
                            temp_dict['doc_links'].append(beginning+link_doc+end)
                    
                    if len(temp_dict['docs']) != 0:
                        data.append(temp_dict)

            await page.go_back()
    
    all_data.extend(data)

Just a note, it's bad that I hardcoded looping through the first 8 pages. Making this auto-updating would require fixing that.

# Separating organizations from document names

orgs = []
loose_orgs = []

for i in range(len(all_data)):
    small_orgs = []
    for j in range(len(all_data[i]['docs'])):
        if all_data[i]['docs'][j].find('Letter') != -1:
            org = all_data[i]['docs'][j].split('Letter')[0].strip()
            small_orgs.append(org)
            loose_orgs.append(org)
        elif all_data[i]['docs'][j].find('Support') != -1:
            org = all_data[i]['docs'][j].split('Support')[0].strip()
            small_orgs.append(org)
            loose_orgs.append(org)
        elif all_data[i]['docs'][j].find('Statement') != -1:
            org = all_data[i]['docs'][j].split('Statement')[0].strip()
            small_orgs.append(org)
            loose_orgs.append(org)
        else:
            pass
        
    orgs.append(small_orgs)

# Separating districts from the titles

districts = []

for i in range(len(all_data)):
    district = str(all_data[i]['position']).lower()
    district = district.split('for the')[-1]
    districts.append(district)


# Making and prepping a dataframe

df = pd.DataFrame(all_data)

df['orgs'] = orgs
df['district'] = districts



pd.to_datetime(df['date'])

df.head()

# Looks alright? I'm going to save it as a csv so avoid scraping it again a bunch of times

df.to_csv('../data/judiciary.csv', index=False)
