==> notebooks/step_2_prep.ipynb <==
# Imports

import requests
from bs4 import BeautifulSoup
import re
from playwright.async_api import async_playwright
import pandas as pd
import time
pd.options.display.max_rows = None
pd.options.display.max_colwidth = None

df = pd.read_csv('../data/judiciary.csv')

df.head()

# So! At this point I want to get the elements of the orgs list into their own frame. And saving the dataframe as a file has turned the lists into strings, so I need to turn them back into lists.

for i in range(len(df)):
    df['orgs'][i] = df['orgs'][i].split(',')
    df['orgs'][i] = [x.strip() for x in df['orgs'][i]]

for i in range(len(df)):
    df['orgs'][i] = [x.replace('[','') for x in df['orgs'][i]]
    df['orgs'][i] = [x.replace(']','') for x in df['orgs'][i]]

for i in range(len(df)):
    df['docs'][i] = df['docs'][i].split(',')
    df['docs'][i] = [x.strip() for x in df['docs'][i]]

for i in range(len(df)):
    df['docs'][i] = [x.replace('[','') for x in df['docs'][i]]
    df['docs'][i] = [x.replace(']','') for x in df['docs'][i]]

for i in range(len(df)):
    df['doc_links'][i] = df['doc_links'][i].split(',')
    df['doc_links'][i] = [x.strip() for x in df['doc_links'][i]]

for i in range(len(df)):
    df['doc_links'][i] = [x.replace('[','') for x in df['doc_links'][i]]
    df['doc_links'][i] = [x.replace(']','') for x in df['doc_links'][i]]

df.head()

# Almost forgot.

pd.to_datetime(df['date'])

# Lovely. Now I want to get the orgs into their own dataframe.

org_list = []

for i in range(len(df)):
    for org in df['orgs'][i]:
       # if org != '':
        temp_dict = {
            'org': org.replace("'",''),
            'name': df['name'][i],
            'position': df['position'][i],
            'date': df['date'][i],
            'district': df['district'][i]
        }
        temp_df = pd.DataFrame(temp_dict, index=[0])
        org_list.append(temp_df)

org_df = pd.concat(org_list, ignore_index=True)
org_df.head()

org_df.org.value_counts()

# Would like to remove blanks, if they aren't mistakes.

names = org_df[org_df.org == ''].name
df[df.name.isin(names)]

# These are just Q&A docs, so I'm going to drop them

org_df = org_df[org_df.org != '']

org_df.org.value_counts()

# Lovely. 

org_df.to_csv('../output/orgs.csv', index=False)
