==> notebooks/step_4_map_prep.ipynb <==
# Imports

import requests
from bs4 import BeautifulSoup
import re
from playwright.async_api import async_playwright
import pandas as pd
import time
import json
import numpy as np
from pandas import json_normalize

# Shamelessly using mapbox tutorial code

with open('../map_files/US_JD_simply.json') as json_data:
    geometry_data = json.load(json_data)



# Reading a dataframe from it

districts_df = pd.DataFrame.from_dict(json_normalize(geometry_data['features']), orient='columns')

districts_df.head()

# reading the orgs dataframe from csv.

orgs_df = pd.read_csv('../output/orgs.csv')


# The dataframe of organizations is way longer than the number of districts! Making a list of dictionaries with information to smush into a dataframe.

districts = []

districts_index = orgs_df.groupby('district').name.count().index

for district in districts_index:
    districts.append(district)


orgs = []

for district in districts:
    orgs.append(orgs_df[orgs_df.district == district].groupby('district').name.count().values[0])

top_judges = []


for district in districts:

      top_judges.append(orgs_df[orgs_df.district == district].groupby('district').name.value_counts().index[0][1])

most_orgs = []

for district in districts:
    most_orgs.append(orgs_df[orgs_df.district == district].groupby('district').name.value_counts().values[0])

big_list = []

for i in range(0, len(districts)):
    temp_dict = {
        'district': districts[i],
        'total_orgs': orgs[i],
        'top_judge': top_judges[i],
        'top_judge_orgs': most_orgs[i]
    }
    big_list.append(temp_dict)

big_list[0]


df = pd.DataFrame(big_list)

# Formatting.

df.district = df.district.str.title().str.strip()
df.district = df.district.str.title().str.strip()
df.district = df.district.str.replace(' Of ', ' of ')

df.head()

# I need a field to merge the dataframes on.

districts_df['match'] = districts_df['properties.DISTRICT'].str.lower()
df['match'] = df.district.str.lower().str.split('district of').str[-1] + " " + df.district.str.lower().str.split('district of').str[0]

df.match = df.match.str.strip()

# The only district (besides some territories that had no data anyway) that this doesn't work for is the District of Columbia, so I have to fix it manually.

for i in range(len(df.match)):
    if df.match[i] == 'columbia':
        df.match[i] = 'district of columbia'
df.match[0:5]

# Merging!

merged_df = pd.merge(districts_df, df, left_on='match', right_on='match', how='left')
merged_df.head()

# Defining and using a function to make the headline property.

def add_text(cell): 
    return "This is in district " + cell

merged_df['properties.headline'] = merged_df['properties.District_N'].apply(add_text)

# Setting the article properties.

merged_df['properties.article'] = ''

for i in range(0, len(merged_df)):
    
    district = str(merged_df['district'][i])
    total_orgs = str(merged_df['total_orgs'][i])
    top_judge = str(merged_df['top_judge'][i])        
    top_judge_orgs = str(merged_df['top_judge_orgs'][i])
    if district != 'nan':
        merged_df['properties.article'][i] = district + ' has had ' + str(int(float(total_orgs))) + ' organizations support a federal judge in the last two years. ' + top_judge + ' had the most with ' + str(int(float(top_judge_orgs))) + '.'
    else:
        merged_df['properties.article'][i] = 'No one cares about federal judges in this district! At least, no one has submitted a brief saying so in the last two years.'


    

# Giving each district a color based on the number of total orgs it has.

# This is less than ideal, I prompted copilot to get the general idea because I'm not super familiar with hex codes.
# The outlier in Washington throws it off, and even after making it a static color the lighter end of the range is too white.

def color(district):
    # get the total orgs for the district
    try:
        total_orgs = merged_df[merged_df.district == district].total_orgs.values[0]

        # These mostly fall between 1 and 50, so we'll use that range.
        # We'll use a linear gradient from light blue to dark blue.
        if total_orgs > 50:
            color = '#%02x%02x%02x' % (0, 0, 255)
        else:
            color = '#%02x%02x%02x' % (int(255 - (total_orgs * 5)*1.1), int(255 - (total_orgs * 5)*1.1), 255)

    except:
        # Generate a gray color if there is no data for the district.
        color = '#%02x%02x%02x' % (128, 128, 128)
    return color

merged_df['properties.color'] = merged_df['district'].apply(color)





# Defining and running functions to make the group dropdown work.

def dis_num(district):
    if district == 'District of Columbia':
        return '12'
    else:
        return district

def dis_name(district):
    if district == 'District of Columbia':
        return district
    else:
        return "District " + district

merged_df['properties.group_id'] = merged_df['properties.District_N'].apply(dis_num)
merged_df['properties.group_name'] = merged_df['properties.District_N'].apply(dis_name)

# All done! Json time.

ok_json = json.loads(merged_df.to_json(orient='records'))

def process_to_geojson(file):
    geo_data = {"type": "FeatureCollection", "features":[]}
    for row in file:
        this_dict = {"type": "Feature", "properties":{}, "geometry": {}}
        for key, value in row.items():
            key_names = key.split('.')
            if key_names[0] == 'geometry':
                this_dict['geometry'][key_names[1]] = value
            if str(key_names[0]) == 'properties':
                this_dict['properties'][key_names[1]] = value
        geo_data['features'].append(this_dict)
    return geo_data

geo_format = process_to_geojson(ok_json)

# variable name
with open('../output/geo-data.js', 'w') as outfile:
    outfile.write("var infoData = ")
# geojson output
with open('../output/geo-data.js', 'a') as outfile:
    json.dump(geo_format, outfile)
